+++ Sources/Superstring/AString.swift
@@ -5,21 +5,11 @@
 //  Created by Manuel Lopes on 03.01.20.
 //
 
-import Foundation
-
 #if canImport(UIKit)
 import UIKit
 #endif
 
-public struct AString: AttributedStringBuildable {
-
-    public var components: AttributedStringComponents {
-        (string, attributes)
-    }
-
-    public var attributedString: NSAttributedString {
-        NSAttributedString(string: components.string, attributes: components.attributes)
-    }
+public struct AString: Equatable, AttributedStringBuildable {
 
     private let string: String
     private let attributes: Attributes
@@ -29,7 +19,15 @@ public struct AString: AttributedStringBuildable {
         self.attributes = attributes
     }
 
-    // MARK: - Internal
+    // MARK: - AttributedStringBuildable
+
+    public var components: AttributedStringComponents {
+        (string, attributes)
+    }
+
+    public var attributedString: NSAttributedString {
+        NSAttributedString(string: components.string, attributes: components.attributes)
+    }
 
     public func applying(_ newAttributes: Attributes) -> Self {
         var attributes = self.attributes
@@ -37,4 +35,11 @@ public struct AString: AttributedStringBuildable {
         attributes.merge(newAttributes, uniquingKeysWith: { (_, new) in new })
         return AString(string, attributes: attributes)
     }
+
+    // MARK: - Equatable
+
+    public static func == (lhs: AString, rhs: AString) -> Bool {
+        lhs.attributedString.isEqual(to: rhs.attributedString)
+    }
+
 }