+++ Tests/SuperstringTests/AStringTests.swift
@@ -19,7 +19,7 @@ final class AStringTests: XCTestCase {
         let expected = NSAttributedString(string: "Superstring")
         let result = AString("Superstring")
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_AString_with_red_background_color() {
@@ -27,7 +27,7 @@ final class AStringTests: XCTestCase {
                                           attributes: [.backgroundColor: Color.red])
         let result = AString("Hello Superstring").backgroundColor(.red)
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_AString_with_multiple_attributes() {
@@ -39,7 +39,7 @@ final class AStringTests: XCTestCase {
             .backgroundColor(.red)
             .foregroundColor(.cyan)
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     static var allTests = [