+++ Tests/SuperstringTests/NewlineTests.swift
@@ -19,7 +19,7 @@ final class NewlineTests: XCTestCase {
         let expected = NSAttributedString(string: "\n")
         let result = Newline()
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_Newline_with_font_attribute() {
@@ -28,7 +28,7 @@ final class NewlineTests: XCTestCase {
         let result = Newline()
             .font(UIFont.systemFont(ofSize: 72))
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     static var allTests = [