+++ Tests/SuperstringTests/SuperstringTests.swift
@@ -14,7 +14,7 @@ final class SuperstringTests: XCTestCase {
             AString("Hello Superstring")
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_super_string_with_one_attribute() {
@@ -25,7 +25,7 @@ final class SuperstringTests: XCTestCase {
                 .backgroundColor(.red)
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_super_string_with_two_AString() {
@@ -40,7 +40,7 @@ final class SuperstringTests: XCTestCase {
             AString("I looked at clouds")
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_super_string_with_newline_and_two_AStrings() {
@@ -56,7 +56,7 @@ final class SuperstringTests: XCTestCase {
             AString("I looked at clouds")
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_super_string_ending_with_Newline() {
@@ -69,7 +69,7 @@ final class SuperstringTests: XCTestCase {
             Newline()
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     func test_super_string_with_single_Newline() {
@@ -80,7 +80,7 @@ final class SuperstringTests: XCTestCase {
             .font(UIFont.systemFont(ofSize: 72))
         }
 
-        XCTAssertTrue(result.attributedString.isEqual(expected))
+        XCTAssertEqual(expected, result.attributedString)
     }
 
     static var allTests = [